+++ Tests/Objectively/IndexPath.c
@@ -27,7 +27,7 @@
 
 START_TEST(indexPath)
 	{
-		int indexes[] = { 1, 2, 3 };
+		size_t indexes[] = { 1, 2, 3 };
 		IndexPath *indexPath = $(alloc(IndexPath), initWithIndexes, indexes, lengthof(indexes));
 
 		ck_assert(indexPath != NULL);
